test(app): move the screen id and plugin routing tests beside their code
`screen_ids_are_unique` and `plugin_id_for_screen_resolves_analytics` test
items that now live in ainb-app and need nothing from the renderer.
